+++ CHANGELOG.md
@@ -3,6 +3,10 @@ _Release Date: Unreleased_
 
 Changelog will appear here when changes are merged.
 
+### 0.3.1
+
+- Fix `/state/non-fungible/data` to return data for all NFIDs
+
 ## 0.3.0 - RCNet-V1
 _Release Date: 31.03.2023_